@@ -1,7 +1,7 @@
 <?php
 /**
  * A helper file for Laravel 5, to provide autocomplete information to your IDE
- * Generated for Laravel 6.6.0 on 2019-11-28.
+ * Generated for Laravel 6.5.2 on 2019-12-03.
  *
  * @author Barry vd. Heuvel <barryvdh@gmail.com>
  * @see https://github.com/barryvdh/laravel-ide-helper
@@ -470,7 +470,7 @@ namespace Illuminate\Support\Facades {
          * @return mixed 
          * @static 
          */ 
-        public static function make($abstract, $parameters = array())
+        public static function make($abstract, $parameters = [])
         {
             return \Illuminate\Foundation\Application::make($abstract, $parameters);
         }
@@ -665,7 +665,7 @@ namespace Illuminate\Support\Facades {
          * @throws \Symfony\Component\HttpKernel\Exception\HttpException
          * @static 
          */ 
-        public static function abort($code, $message = '', $headers = array())
+        public static function abort($code, $message = '', $headers = [])
         {
             \Illuminate\Foundation\Application::abort($code, $message, $headers);
         }
@@ -1124,7 +1124,7 @@ namespace Illuminate\Support\Facades {
          * @return \Closure 
          * @static 
          */ 
-        public static function wrap($callback, $parameters = array())
+        public static function wrap($callback, $parameters = [])
         {
             //Method inherited from \Illuminate\Container\Container            
             return \Illuminate\Foundation\Application::wrap($callback, $parameters);
@@ -1139,7 +1139,7 @@ namespace Illuminate\Support\Facades {
          * @return mixed 
          * @static 
          */ 
-        public static function call($callback, $parameters = array(), $defaultMethod = null)
+        public static function call($callback, $parameters = [], $defaultMethod = null)
         {
             //Method inherited from \Illuminate\Container\Container            
             return \Illuminate\Foundation\Application::call($callback, $parameters, $defaultMethod);
@@ -1166,7 +1166,7 @@ namespace Illuminate\Support\Facades {
          * @return mixed 
          * @static 
          */ 
-        public static function makeWith($abstract, $parameters = array())
+        public static function makeWith($abstract, $parameters = [])
         {
             //Method inherited from \Illuminate\Container\Container            
             return \Illuminate\Foundation\Application::makeWith($abstract, $parameters);
@@ -1439,7 +1439,7 @@ namespace Illuminate\Support\Facades {
          * @throws \Symfony\Component\Console\Exception\CommandNotFoundException
          * @static 
          */ 
-        public static function call($command, $parameters = array(), $outputBuffer = null)
+        public static function call($command, $parameters = [], $outputBuffer = null)
         {
             //Method inherited from \Illuminate\Foundation\Console\Kernel            
             return \App\Console\Kernel::call($command, $parameters, $outputBuffer);
@@ -1453,7 +1453,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Foundation\Bus\PendingDispatch 
          * @static 
          */ 
-        public static function queue($command, $parameters = array())
+        public static function queue($command, $parameters = [])
         {
             //Method inherited from \Illuminate\Foundation\Console\Kernel            
             return \App\Console\Kernel::queue($command, $parameters);
@@ -1711,7 +1711,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function once($credentials = array())
+        public static function once($credentials = [])
         {
             return \Illuminate\Auth\SessionGuard::once($credentials);
         }
@@ -1735,7 +1735,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function validate($credentials = array())
+        public static function validate($credentials = [])
         {
             return \Illuminate\Auth\SessionGuard::validate($credentials);
         }
@@ -1748,7 +1748,7 @@ namespace Illuminate\Support\Facades {
          * @return \Symfony\Component\HttpFoundation\Response|null 
          * @static 
          */ 
-        public static function basic($field = 'email', $extraConditions = array())
+        public static function basic($field = 'email', $extraConditions = [])
         {
             return \Illuminate\Auth\SessionGuard::basic($field, $extraConditions);
         }
@@ -1761,7 +1761,7 @@ namespace Illuminate\Support\Facades {
          * @return \Symfony\Component\HttpFoundation\Response|null 
          * @static 
          */ 
-        public static function onceBasic($field = 'email', $extraConditions = array())
+        public static function onceBasic($field = 'email', $extraConditions = [])
         {
             return \Illuminate\Auth\SessionGuard::onceBasic($field, $extraConditions);
         }
@@ -1774,7 +1774,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function attempt($credentials = array(), $remember = false)
+        public static function attempt($credentials = [], $remember = false)
         {
             return \Illuminate\Auth\SessionGuard::attempt($credentials, $remember);
         }
@@ -3814,7 +3814,7 @@ namespace Illuminate\Support\Facades {
          * @return mixed 
          * @static 
          */ 
-        public static function selectOne($query, $bindings = array(), $useReadPdo = true)
+        public static function selectOne($query, $bindings = [], $useReadPdo = true)
         {
             //Method inherited from \Illuminate\Database\Connection            
             return \Illuminate\Database\MySqlConnection::selectOne($query, $bindings, $useReadPdo);
@@ -3828,7 +3828,7 @@ namespace Illuminate\Support\Facades {
          * @return array 
          * @static 
          */ 
-        public static function selectFromWriteConnection($query, $bindings = array())
+        public static function selectFromWriteConnection($query, $bindings = [])
         {
             //Method inherited from \Illuminate\Database\Connection            
             return \Illuminate\Database\MySqlConnection::selectFromWriteConnection($query, $bindings);
@@ -3843,7 +3843,7 @@ namespace Illuminate\Support\Facades {
          * @return array 
          * @static 
          */ 
-        public static function select($query, $bindings = array(), $useReadPdo = true)
+        public static function select($query, $bindings = [], $useReadPdo = true)
         {
             //Method inherited from \Illuminate\Database\Connection            
             return \Illuminate\Database\MySqlConnection::select($query, $bindings, $useReadPdo);
@@ -3858,7 +3858,7 @@ namespace Illuminate\Support\Facades {
          * @return \Generator 
          * @static 
          */ 
-        public static function cursor($query, $bindings = array(), $useReadPdo = true)
+        public static function cursor($query, $bindings = [], $useReadPdo = true)
         {
             //Method inherited from \Illuminate\Database\Connection            
             return \Illuminate\Database\MySqlConnection::cursor($query, $bindings, $useReadPdo);
@@ -3872,7 +3872,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function insert($query, $bindings = array())
+        public static function insert($query, $bindings = [])
         {
             //Method inherited from \Illuminate\Database\Connection            
             return \Illuminate\Database\MySqlConnection::insert($query, $bindings);
@@ -3886,7 +3886,7 @@ namespace Illuminate\Support\Facades {
          * @return int 
          * @static 
          */ 
-        public static function update($query, $bindings = array())
+        public static function update($query, $bindings = [])
         {
             //Method inherited from \Illuminate\Database\Connection            
             return \Illuminate\Database\MySqlConnection::update($query, $bindings);
@@ -3900,7 +3900,7 @@ namespace Illuminate\Support\Facades {
          * @return int 
          * @static 
          */ 
-        public static function delete($query, $bindings = array())
+        public static function delete($query, $bindings = [])
         {
             //Method inherited from \Illuminate\Database\Connection            
             return \Illuminate\Database\MySqlConnection::delete($query, $bindings);
@@ -3914,7 +3914,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function statement($query, $bindings = array())
+        public static function statement($query, $bindings = [])
         {
             //Method inherited from \Illuminate\Database\Connection            
             return \Illuminate\Database\MySqlConnection::statement($query, $bindings);
@@ -3928,7 +3928,7 @@ namespace Illuminate\Support\Facades {
          * @return int 
          * @static 
          */ 
-        public static function affectingStatement($query, $bindings = array())
+        public static function affectingStatement($query, $bindings = [])
         {
             //Method inherited from \Illuminate\Database\Connection            
             return \Illuminate\Database\MySqlConnection::affectingStatement($query, $bindings);
@@ -4541,7 +4541,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function push($event, $payload = array())
+        public static function push($event, $payload = [])
         {
             \Illuminate\Events\Dispatcher::push($event, $payload);
         }
@@ -4578,7 +4578,7 @@ namespace Illuminate\Support\Facades {
          * @return array|null 
          * @static 
          */ 
-        public static function until($event, $payload = array())
+        public static function until($event, $payload = [])
         {
             return \Illuminate\Events\Dispatcher::until($event, $payload);
         }
@@ -4592,7 +4592,7 @@ namespace Illuminate\Support\Facades {
          * @return array|null 
          * @static 
          */ 
-        public static function dispatch($event, $payload = array(), $halt = false)
+        public static function dispatch($event, $payload = [], $halt = false)
         {
             return \Illuminate\Events\Dispatcher::dispatch($event, $payload, $halt);
         }
@@ -5347,7 +5347,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function allows($ability, $arguments = array())
+        public static function allows($ability, $arguments = [])
         {
             return \Illuminate\Auth\Access\Gate::allows($ability, $arguments);
         }
@@ -5360,7 +5360,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function denies($ability, $arguments = array())
+        public static function denies($ability, $arguments = [])
         {
             return \Illuminate\Auth\Access\Gate::denies($ability, $arguments);
         }
@@ -5373,7 +5373,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function check($abilities, $arguments = array())
+        public static function check($abilities, $arguments = [])
         {
             return \Illuminate\Auth\Access\Gate::check($abilities, $arguments);
         }
@@ -5386,7 +5386,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function any($abilities, $arguments = array())
+        public static function any($abilities, $arguments = [])
         {
             return \Illuminate\Auth\Access\Gate::any($abilities, $arguments);
         }
@@ -5399,7 +5399,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function none($abilities, $arguments = array())
+        public static function none($abilities, $arguments = [])
         {
             return \Illuminate\Auth\Access\Gate::none($abilities, $arguments);
         }
@@ -5413,7 +5413,7 @@ namespace Illuminate\Support\Facades {
          * @throws \Illuminate\Auth\Access\AuthorizationException
          * @static 
          */ 
-        public static function authorize($ability, $arguments = array())
+        public static function authorize($ability, $arguments = [])
         {
             return \Illuminate\Auth\Access\Gate::authorize($ability, $arguments);
         }
@@ -5426,7 +5426,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Auth\Access\Response 
          * @static 
          */ 
-        public static function inspect($ability, $arguments = array())
+        public static function inspect($ability, $arguments = [])
         {
             return \Illuminate\Auth\Access\Gate::inspect($ability, $arguments);
         }
@@ -5440,7 +5440,7 @@ namespace Illuminate\Support\Facades {
          * @throws \Illuminate\Auth\Access\AuthorizationException
          * @static 
          */ 
-        public static function raw($ability, $arguments = array())
+        public static function raw($ability, $arguments = [])
         {
             return \Illuminate\Auth\Access\Gate::raw($ability, $arguments);
         }
@@ -5573,7 +5573,7 @@ namespace Illuminate\Support\Facades {
          * @return string 
          * @static 
          */ 
-        public static function make($value, $options = array())
+        public static function make($value, $options = [])
         {
             return \Illuminate\Hashing\HashManager::make($value, $options);
         }
@@ -5587,7 +5587,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function check($value, $hashedValue, $options = array())
+        public static function check($value, $hashedValue, $options = [])
         {
             return \Illuminate\Hashing\HashManager::check($value, $hashedValue, $options);
         }
@@ -5600,7 +5600,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function needsRehash($hashedValue, $options = array())
+        public static function needsRehash($hashedValue, $options = [])
         {
             return \Illuminate\Hashing\HashManager::needsRehash($hashedValue, $options);
         }
@@ -5697,7 +5697,7 @@ namespace Illuminate\Support\Facades {
          * @return string|array 
          * @static 
          */ 
-        public static function get($key, $replace = array(), $locale = null, $fallback = true)
+        public static function get($key, $replace = [], $locale = null, $fallback = true)
         {
             return \Illuminate\Translation\Translator::get($key, $replace, $locale, $fallback);
         }
@@ -5712,7 +5712,7 @@ namespace Illuminate\Support\Facades {
          * @return string 
          * @static 
          */ 
-        public static function choice($key, $number, $replace = array(), $locale = null)
+        public static function choice($key, $number, $replace = [], $locale = null)
         {
             return \Illuminate\Translation\Translator::choice($key, $number, $replace, $locale);
         }
@@ -6046,7 +6046,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function emergency($message, $context = array())
+        public static function emergency($message, $context = [])
         {
             \Illuminate\Log\LogManager::emergency($message, $context);
         }
@@ -6062,7 +6062,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function alert($message, $context = array())
+        public static function alert($message, $context = [])
         {
             \Illuminate\Log\LogManager::alert($message, $context);
         }
@@ -6077,7 +6077,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function critical($message, $context = array())
+        public static function critical($message, $context = [])
         {
             \Illuminate\Log\LogManager::critical($message, $context);
         }
@@ -6091,7 +6091,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function error($message, $context = array())
+        public static function error($message, $context = [])
         {
             \Illuminate\Log\LogManager::error($message, $context);
         }
@@ -6107,7 +6107,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function warning($message, $context = array())
+        public static function warning($message, $context = [])
         {
             \Illuminate\Log\LogManager::warning($message, $context);
         }
@@ -6120,7 +6120,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function notice($message, $context = array())
+        public static function notice($message, $context = [])
         {
             \Illuminate\Log\LogManager::notice($message, $context);
         }
@@ -6135,7 +6135,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function info($message, $context = array())
+        public static function info($message, $context = [])
         {
             \Illuminate\Log\LogManager::info($message, $context);
         }
@@ -6148,7 +6148,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function debug($message, $context = array())
+        public static function debug($message, $context = [])
         {
             \Illuminate\Log\LogManager::debug($message, $context);
         }
@@ -6162,7 +6162,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function log($level, $message, $context = array())
+        public static function log($level, $message, $context = [])
         {
             \Illuminate\Log\LogManager::log($level, $message, $context);
         }
@@ -6294,7 +6294,7 @@ namespace Illuminate\Support\Facades {
          * @return string 
          * @static 
          */ 
-        public static function render($view, $data = array())
+        public static function render($view, $data = [])
         {
             return \Illuminate\Mail\Mailer::render($view, $data);
         }
@@ -6308,7 +6308,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function send($view, $data = array(), $callback = null)
+        public static function send($view, $data = [], $callback = null)
         {
             \Illuminate\Mail\Mailer::send($view, $data, $callback);
         }
@@ -7101,7 +7101,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function assertPushedWithChain($job, $expectedChain = array(), $callback = null)
+        public static function assertPushedWithChain($job, $expectedChain = [], $callback = null)
         {
             \Illuminate\Support\Testing\Fakes\QueueFake::assertPushedWithChain($job, $expectedChain, $callback);
         }
@@ -7190,7 +7190,7 @@ namespace Illuminate\Support\Facades {
          * @return mixed 
          * @static 
          */ 
-        public static function pushRaw($payload, $queue = null, $options = array())
+        public static function pushRaw($payload, $queue = null, $options = [])
         {
             return \Illuminate\Support\Testing\Fakes\QueueFake::pushRaw($payload, $queue, $options);
         }
@@ -7376,7 +7376,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function back($status = 302, $headers = array(), $fallback = false)
+        public static function back($status = 302, $headers = [], $fallback = false)
         {
             return \Illuminate\Routing\Redirector::back($status, $headers, $fallback);
         }
@@ -7389,7 +7389,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function refresh($status = 302, $headers = array())
+        public static function refresh($status = 302, $headers = [])
         {
             return \Illuminate\Routing\Redirector::refresh($status, $headers);
         }
@@ -7404,7 +7404,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function guest($path, $status = 302, $headers = array(), $secure = null)
+        public static function guest($path, $status = 302, $headers = [], $secure = null)
         {
             return \Illuminate\Routing\Redirector::guest($path, $status, $headers, $secure);
         }
@@ -7419,7 +7419,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function intended($default = '/', $status = 302, $headers = array(), $secure = null)
+        public static function intended($default = '/', $status = 302, $headers = [], $secure = null)
         {
             return \Illuminate\Routing\Redirector::intended($default, $status, $headers, $secure);
         }
@@ -7446,7 +7446,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function to($path, $status = 302, $headers = array(), $secure = null)
+        public static function to($path, $status = 302, $headers = [], $secure = null)
         {
             return \Illuminate\Routing\Redirector::to($path, $status, $headers, $secure);
         }
@@ -7460,7 +7460,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function away($path, $status = 302, $headers = array())
+        public static function away($path, $status = 302, $headers = [])
         {
             return \Illuminate\Routing\Redirector::away($path, $status, $headers);
         }
@@ -7474,7 +7474,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function secure($path, $status = 302, $headers = array())
+        public static function secure($path, $status = 302, $headers = [])
         {
             return \Illuminate\Routing\Redirector::secure($path, $status, $headers);
         }
@@ -7489,7 +7489,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function route($route, $parameters = array(), $status = 302, $headers = array())
+        public static function route($route, $parameters = [], $status = 302, $headers = [])
         {
             return \Illuminate\Routing\Redirector::route($route, $parameters, $status, $headers);
         }
@@ -7504,7 +7504,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function action($action, $parameters = array(), $status = 302, $headers = array())
+        public static function action($action, $parameters = [], $status = 302, $headers = [])
         {
             return \Illuminate\Routing\Redirector::action($action, $parameters, $status, $headers);
         }
@@ -7569,6 +7569,93 @@ namespace Illuminate\Support\Facades {
         public static function hasMacro($name)
         {
             return \Illuminate\Routing\Redirector::hasMacro($name);
+        }
+         
+    }
+
+    class Redis {
+        
+        /**
+         * Get a Redis connection by name.
+         *
+         * @param string|null $name
+         * @return \Illuminate\Redis\Connections\Connection 
+         * @static 
+         */ 
+        public static function connection($name = null)
+        {
+            return \Illuminate\Redis\RedisManager::connection($name);
+        }
+        
+        /**
+         * Resolve the given connection by name.
+         *
+         * @param string|null $name
+         * @return \Illuminate\Redis\Connections\Connection 
+         * @throws \InvalidArgumentException
+         * @static 
+         */ 
+        public static function resolve($name = null)
+        {
+            return \Illuminate\Redis\RedisManager::resolve($name);
+        }
+        
+        /**
+         * Return all of the created connections.
+         *
+         * @return array 
+         * @static 
+         */ 
+        public static function connections()
+        {
+            return \Illuminate\Redis\RedisManager::connections();
+        }
+        
+        /**
+         * Enable the firing of Redis command events.
+         *
+         * @return void 
+         * @static 
+         */ 
+        public static function enableEvents()
+        {
+            \Illuminate\Redis\RedisManager::enableEvents();
+        }
+        
+        /**
+         * Disable the firing of Redis command events.
+         *
+         * @return void 
+         * @static 
+         */ 
+        public static function disableEvents()
+        {
+            \Illuminate\Redis\RedisManager::disableEvents();
+        }
+        
+        /**
+         * Set the default driver.
+         *
+         * @param string $driver
+         * @return void 
+         * @static 
+         */ 
+        public static function setDriver($driver)
+        {
+            \Illuminate\Redis\RedisManager::setDriver($driver);
+        }
+        
+        /**
+         * Register a custom driver creator Closure.
+         *
+         * @param string $driver
+         * @param \Closure $callback
+         * @return \Illuminate\Redis\RedisManager 
+         * @static 
+         */ 
+        public static function extend($driver, $callback)
+        {
+            return \Illuminate\Redis\RedisManager::extend($driver, $callback);
         }
          
     }
@@ -8110,7 +8197,7 @@ namespace Illuminate\Support\Facades {
          * @param string|resource|null $content The raw body data
          * @static 
          */ 
-        public static function initialize($query = array(), $request = array(), $attributes = array(), $cookies = array(), $files = array(), $server = array(), $content = null)
+        public static function initialize($query = [], $request = [], $attributes = [], $cookies = [], $files = [], $server = [], $content = null)
         {
             //Method inherited from \Symfony\Component\HttpFoundation\Request            
             return \Illuminate\Http\Request::initialize($query, $request, $attributes, $cookies, $files, $server, $content);
@@ -8144,7 +8231,7 @@ namespace Illuminate\Support\Facades {
          * @return static 
          * @static 
          */ 
-        public static function create($uri, $method = 'GET', $parameters = array(), $cookies = array(), $files = array(), $server = array(), $content = null)
+        public static function create($uri, $method = 'GET', $parameters = [], $cookies = [], $files = [], $server = [], $content = null)
         {
             //Method inherited from \Symfony\Component\HttpFoundation\Request            
             return \Illuminate\Http\Request::create($uri, $method, $parameters, $cookies, $files, $server, $content);
@@ -9581,7 +9668,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\Response 
          * @static 
          */ 
-        public static function make($content = '', $status = 200, $headers = array())
+        public static function make($content = '', $status = 200, $headers = [])
         {
             return \Illuminate\Routing\ResponseFactory::make($content, $status, $headers);
         }
@@ -9594,7 +9681,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\Response 
          * @static 
          */ 
-        public static function noContent($status = 204, $headers = array())
+        public static function noContent($status = 204, $headers = [])
         {
             return \Illuminate\Routing\ResponseFactory::noContent($status, $headers);
         }
@@ -9602,14 +9689,14 @@ namespace Illuminate\Support\Facades {
         /**
          * Create a new response for a given view.
          *
-         * @param string|array $view
+         * @param string $view
          * @param array $data
          * @param int $status
          * @param array $headers
          * @return \Illuminate\Http\Response 
          * @static 
          */ 
-        public static function view($view, $data = array(), $status = 200, $headers = array())
+        public static function view($view, $data = [], $status = 200, $headers = [])
         {
             return \Illuminate\Routing\ResponseFactory::view($view, $data, $status, $headers);
         }
@@ -9624,7 +9711,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\JsonResponse 
          * @static 
          */ 
-        public static function json($data = array(), $status = 200, $headers = array(), $options = 0)
+        public static function json($data = [], $status = 200, $headers = [], $options = 0)
         {
             return \Illuminate\Routing\ResponseFactory::json($data, $status, $headers, $options);
         }
@@ -9640,7 +9727,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\JsonResponse 
          * @static 
          */ 
-        public static function jsonp($callback, $data = array(), $status = 200, $headers = array(), $options = 0)
+        public static function jsonp($callback, $data = [], $status = 200, $headers = [], $options = 0)
         {
             return \Illuminate\Routing\ResponseFactory::jsonp($callback, $data, $status, $headers, $options);
         }
@@ -9654,7 +9741,7 @@ namespace Illuminate\Support\Facades {
          * @return \Symfony\Component\HttpFoundation\StreamedResponse 
          * @static 
          */ 
-        public static function stream($callback, $status = 200, $headers = array())
+        public static function stream($callback, $status = 200, $headers = [])
         {
             return \Illuminate\Routing\ResponseFactory::stream($callback, $status, $headers);
         }
@@ -9669,7 +9756,7 @@ namespace Illuminate\Support\Facades {
          * @return \Symfony\Component\HttpFoundation\StreamedResponse 
          * @static 
          */ 
-        public static function streamDownload($callback, $name = null, $headers = array(), $disposition = 'attachment')
+        public static function streamDownload($callback, $name = null, $headers = [], $disposition = 'attachment')
         {
             return \Illuminate\Routing\ResponseFactory::streamDownload($callback, $name, $headers, $disposition);
         }
@@ -9684,7 +9771,7 @@ namespace Illuminate\Support\Facades {
          * @return \Symfony\Component\HttpFoundation\BinaryFileResponse 
          * @static 
          */ 
-        public static function download($file, $name = null, $headers = array(), $disposition = 'attachment')
+        public static function download($file, $name = null, $headers = [], $disposition = 'attachment')
         {
             return \Illuminate\Routing\ResponseFactory::download($file, $name, $headers, $disposition);
         }
@@ -9697,7 +9784,7 @@ namespace Illuminate\Support\Facades {
          * @return \Symfony\Component\HttpFoundation\BinaryFileResponse 
          * @static 
          */ 
-        public static function file($file, $headers = array())
+        public static function file($file, $headers = [])
         {
             return \Illuminate\Routing\ResponseFactory::file($file, $headers);
         }
@@ -9712,7 +9799,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function redirectTo($path, $status = 302, $headers = array(), $secure = null)
+        public static function redirectTo($path, $status = 302, $headers = [], $secure = null)
         {
             return \Illuminate\Routing\ResponseFactory::redirectTo($path, $status, $headers, $secure);
         }
@@ -9727,7 +9814,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function redirectToRoute($route, $parameters = array(), $status = 302, $headers = array())
+        public static function redirectToRoute($route, $parameters = [], $status = 302, $headers = [])
         {
             return \Illuminate\Routing\ResponseFactory::redirectToRoute($route, $parameters, $status, $headers);
         }
@@ -9742,7 +9829,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function redirectToAction($action, $parameters = array(), $status = 302, $headers = array())
+        public static function redirectToAction($action, $parameters = [], $status = 302, $headers = [])
         {
             return \Illuminate\Routing\ResponseFactory::redirectToAction($action, $parameters, $status, $headers);
         }
@@ -9757,7 +9844,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function redirectGuest($path, $status = 302, $headers = array(), $secure = null)
+        public static function redirectGuest($path, $status = 302, $headers = [], $secure = null)
         {
             return \Illuminate\Routing\ResponseFactory::redirectGuest($path, $status, $headers, $secure);
         }
@@ -9772,7 +9859,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Http\RedirectResponse 
          * @static 
          */ 
-        public static function redirectToIntended($default = '/', $status = 302, $headers = array(), $secure = null)
+        public static function redirectToIntended($default = '/', $status = 302, $headers = [], $secure = null)
         {
             return \Illuminate\Routing\ResponseFactory::redirectToIntended($default, $status, $headers, $secure);
         }
@@ -9819,19 +9906,6 @@ namespace Illuminate\Support\Facades {
     }
 
     class Route {
-        
-        /**
-         * Register a new HEAD route with the router.
-         *
-         * @param string $uri
-         * @param \Closure|array|string|callable|null $action
-         * @return \Illuminate\Routing\Route 
-         * @static 
-         */ 
-        public static function head($uri, $action = null)
-        {
-            return \Illuminate\Routing\Router::head($uri, $action);
-        }
         
         /**
          * Register a new GET route with the router.
@@ -9972,7 +10046,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Routing\Route 
          * @static 
          */ 
-        public static function view($uri, $view, $data = array())
+        public static function view($uri, $view, $data = [])
         {
             return \Illuminate\Routing\Router::view($uri, $view, $data);
         }
@@ -9999,7 +10073,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function resources($resources, $options = array())
+        public static function resources($resources, $options = [])
         {
             \Illuminate\Routing\Router::resources($resources, $options);
         }
@@ -10013,7 +10087,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Routing\PendingResourceRegistration 
          * @static 
          */ 
-        public static function resource($name, $controller, $options = array())
+        public static function resource($name, $controller, $options = [])
         {
             return \Illuminate\Routing\Router::resource($name, $controller, $options);
         }
@@ -10026,7 +10100,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function apiResources($resources, $options = array())
+        public static function apiResources($resources, $options = [])
         {
             \Illuminate\Routing\Router::apiResources($resources, $options);
         }
@@ -10040,7 +10114,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Routing\PendingResourceRegistration 
          * @static 
          */ 
-        public static function apiResource($name, $controller, $options = array())
+        public static function apiResource($name, $controller, $options = [])
         {
             return \Illuminate\Routing\Router::apiResource($name, $controller, $options);
         }
@@ -10529,7 +10603,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function auth($options = array())
+        public static function auth($options = [])
         {
             \Illuminate\Routing\Router::auth($options);
         }
@@ -10586,7 +10660,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function resourceParameters($parameters = array())
+        public static function resourceParameters($parameters = [])
         {
             \Illuminate\Routing\Router::resourceParameters($parameters);
         }
@@ -10598,7 +10672,7 @@ namespace Illuminate\Support\Facades {
          * @return array|null 
          * @static 
          */ 
-        public static function resourceVerbs($verbs = array())
+        public static function resourceVerbs($verbs = [])
         {
             return \Illuminate\Routing\Router::resourceVerbs($verbs);
         }
@@ -11761,7 +11835,7 @@ namespace Illuminate\Support\Facades {
          * @return \Symfony\Component\HttpFoundation\StreamedResponse 
          * @static 
          */ 
-        public static function response($path, $name = null, $headers = array(), $disposition = 'inline')
+        public static function response($path, $name = null, $headers = [], $disposition = 'inline')
         {
             return \Illuminate\Filesystem\FilesystemAdapter::response($path, $name, $headers, $disposition);
         }
@@ -11775,7 +11849,7 @@ namespace Illuminate\Support\Facades {
          * @return \Symfony\Component\HttpFoundation\StreamedResponse 
          * @static 
          */ 
-        public static function download($path, $name = null, $headers = array())
+        public static function download($path, $name = null, $headers = [])
         {
             return \Illuminate\Filesystem\FilesystemAdapter::download($path, $name, $headers);
         }
@@ -11789,7 +11863,7 @@ namespace Illuminate\Support\Facades {
          * @return bool 
          * @static 
          */ 
-        public static function put($path, $contents, $options = array())
+        public static function put($path, $contents, $options = [])
         {
             return \Illuminate\Filesystem\FilesystemAdapter::put($path, $contents, $options);
         }
@@ -11803,7 +11877,7 @@ namespace Illuminate\Support\Facades {
          * @return string|false 
          * @static 
          */ 
-        public static function putFile($path, $file, $options = array())
+        public static function putFile($path, $file, $options = [])
         {
             return \Illuminate\Filesystem\FilesystemAdapter::putFile($path, $file, $options);
         }
@@ -11818,7 +11892,7 @@ namespace Illuminate\Support\Facades {
          * @return string|false 
          * @static 
          */ 
-        public static function putFileAs($path, $file, $name, $options = array())
+        public static function putFileAs($path, $file, $name, $options = [])
         {
             return \Illuminate\Filesystem\FilesystemAdapter::putFileAs($path, $file, $name, $options);
         }
@@ -11987,7 +12061,7 @@ namespace Illuminate\Support\Facades {
          * @throws \Illuminate\Contracts\Filesystem\FileExistsException
          * @static 
          */ 
-        public static function writeStream($path, $resource, $options = array())
+        public static function writeStream($path, $resource, $options = [])
         {
             return \Illuminate\Filesystem\FilesystemAdapter::writeStream($path, $resource, $options);
         }
@@ -12002,7 +12076,7 @@ namespace Illuminate\Support\Facades {
          * @throws \RuntimeException
          * @static 
          */ 
-        public static function temporaryUrl($path, $expiration, $options = array())
+        public static function temporaryUrl($path, $expiration, $options = [])
         {
             return \Illuminate\Filesystem\FilesystemAdapter::temporaryUrl($path, $expiration, $options);
         }
@@ -12165,7 +12239,7 @@ namespace Illuminate\Support\Facades {
          * @return string 
          * @static 
          */ 
-        public static function to($path, $extra = array(), $secure = null)
+        public static function to($path, $extra = [], $secure = null)
         {
             return \Illuminate\Routing\UrlGenerator::to($path, $extra, $secure);
         }
@@ -12178,7 +12252,7 @@ namespace Illuminate\Support\Facades {
          * @return string 
          * @static 
          */ 
-        public static function secure($path, $parameters = array())
+        public static function secure($path, $parameters = [])
         {
             return \Illuminate\Routing\UrlGenerator::secure($path, $parameters);
         }
@@ -12245,7 +12319,7 @@ namespace Illuminate\Support\Facades {
          * @throws \InvalidArgumentException
          * @static 
          */ 
-        public static function signedRoute($name, $parameters = array(), $expiration = null, $absolute = true)
+        public static function signedRoute($name, $parameters = [], $expiration = null, $absolute = true)
         {
             return \Illuminate\Routing\UrlGenerator::signedRoute($name, $parameters, $expiration, $absolute);
         }
@@ -12260,7 +12334,7 @@ namespace Illuminate\Support\Facades {
          * @return string 
          * @static 
          */ 
-        public static function temporarySignedRoute($name, $expiration, $parameters = array(), $absolute = true)
+        public static function temporarySignedRoute($name, $expiration, $parameters = [], $absolute = true)
         {
             return \Illuminate\Routing\UrlGenerator::temporarySignedRoute($name, $expiration, $parameters, $absolute);
         }
@@ -12313,7 +12387,7 @@ namespace Illuminate\Support\Facades {
          * @throws \Symfony\Component\Routing\Exception\RouteNotFoundException
          * @static 
          */ 
-        public static function route($name, $parameters = array(), $absolute = true)
+        public static function route($name, $parameters = [], $absolute = true)
         {
             return \Illuminate\Routing\UrlGenerator::route($name, $parameters, $absolute);
         }
@@ -12343,7 +12417,7 @@ namespace Illuminate\Support\Facades {
          * @throws \InvalidArgumentException
          * @static 
          */ 
-        public static function action($action, $parameters = array(), $absolute = true)
+        public static function action($action, $parameters = [], $absolute = true)
         {
             return \Illuminate\Routing\UrlGenerator::action($action, $parameters, $absolute);
         }
@@ -12605,7 +12679,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Validation\Validator 
          * @static 
          */ 
-        public static function make($data, $rules, $messages = array(), $customAttributes = array())
+        public static function make($data, $rules, $messages = [], $customAttributes = [])
         {
             return \Illuminate\Validation\Factory::make($data, $rules, $messages, $customAttributes);
         }
@@ -12621,7 +12695,7 @@ namespace Illuminate\Support\Facades {
          * @throws \Illuminate\Validation\ValidationException
          * @static 
          */ 
-        public static function validate($data, $rules, $messages = array(), $customAttributes = array())
+        public static function validate($data, $rules, $messages = [], $customAttributes = [])
         {
             return \Illuminate\Validation\Factory::validate($data, $rules, $messages, $customAttributes);
         }
@@ -12740,7 +12814,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Contracts\View\View 
          * @static 
          */ 
-        public static function file($path, $data = array(), $mergeData = array())
+        public static function file($path, $data = [], $mergeData = [])
         {
             return \Illuminate\View\Factory::file($path, $data, $mergeData);
         }
@@ -12754,7 +12828,7 @@ namespace Illuminate\Support\Facades {
          * @return \Illuminate\Contracts\View\View 
          * @static 
          */ 
-        public static function make($view, $data = array(), $mergeData = array())
+        public static function make($view, $data = [], $mergeData = [])
         {
             return \Illuminate\View\Factory::make($view, $data, $mergeData);
         }
@@ -12769,7 +12843,7 @@ namespace Illuminate\Support\Facades {
          * @throws \InvalidArgumentException
          * @static 
          */ 
-        public static function first($views, $data = array(), $mergeData = array())
+        public static function first($views, $data = [], $mergeData = [])
         {
             return \Illuminate\View\Factory::first($views, $data, $mergeData);
         }
@@ -12784,7 +12858,7 @@ namespace Illuminate\Support\Facades {
          * @return string 
          * @static 
          */ 
-        public static function renderWhen($condition, $view, $data = array(), $mergeData = array())
+        public static function renderWhen($condition, $view, $data = [], $mergeData = [])
         {
             return \Illuminate\View\Factory::renderWhen($condition, $view, $data, $mergeData);
         }
@@ -13135,7 +13209,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function startComponent($name, $data = array())
+        public static function startComponent($name, $data = [])
         {
             \Illuminate\View\Factory::startComponent($name, $data);
         }
@@ -13148,7 +13222,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function startComponentFirst($names, $data = array())
+        public static function startComponentFirst($names, $data = [])
         {
             \Illuminate\View\Factory::startComponentFirst($names, $data);
         }
@@ -13521,7 +13595,7 @@ namespace Illuminate\Support\Facades {
          * @return void 
          * @static 
          */ 
-        public static function startTranslation($replacements = array())
+        public static function startTranslation($replacements = [])
         {
             \Illuminate\View\Factory::startTranslation($replacements);
         }
@@ -13564,7 +13638,7 @@ namespace Intervention\Image\Facades {
          * @return self 
          * @static 
          */ 
-        public static function configure($config = array())
+        public static function configure($config = [])
         {
             return \Intervention\Image\ImageManager::configure($config);
         }
@@ -13688,7 +13762,7 @@ namespace Barryvdh\Debugbar {
          * @throws \ErrorException
          * @static 
          */ 
-        public static function handleError($level, $message, $file = '', $line = 0, $context = array())
+        public static function handleError($level, $message, $file = '', $line = 0, $context = [])
         {
             return \Barryvdh\Debugbar\LaravelDebugbar::handleError($level, $message, $file, $line, $context);
         }
@@ -14472,7 +14546,7 @@ namespace Facade\Ignition\Facades {
          *
          * @static 
          */ 
-        public static function glow($name, $messageLevel = 'info', $metaData = array())
+        public static function glow($name, $messageLevel = 'info', $metaData = [])
         {
             return \Facade\FlareClient\Flare::glow($name, $messageLevel, $metaData);
         }
@@ -14602,7 +14676,7 @@ namespace Facade\Ignition\Facades {
          *
          * @static 
          */ 
-        public static function getGroup($groupName = 'context', $default = array())
+        public static function getGroup($groupName = 'context', $default = [])
         {
             return \Facade\FlareClient\Flare::getGroup($groupName, $default);
         }
@@ -14642,7 +14716,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function open($options = array())
+        public static function open($options = [])
         {
             return \Collective\Html\FormBuilder::open($options);
         }
@@ -14655,7 +14729,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function model($model, $options = array())
+        public static function model($model, $options = [])
         {
             return \Collective\Html\FormBuilder::model($model, $options);
         }
@@ -14715,7 +14789,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function label($name, $value = null, $options = array(), $escape_html = true)
+        public static function label($name, $value = null, $options = [], $escape_html = true)
         {
             return \Collective\Html\FormBuilder::label($name, $value, $options, $escape_html);
         }
@@ -14730,7 +14804,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function input($type, $name, $value = null, $options = array())
+        public static function input($type, $name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::input($type, $name, $value, $options);
         }
@@ -14744,7 +14818,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function text($name, $value = null, $options = array())
+        public static function text($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::text($name, $value, $options);
         }
@@ -14757,7 +14831,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function password($name, $options = array())
+        public static function password($name, $options = [])
         {
             return \Collective\Html\FormBuilder::password($name, $options);
         }
@@ -14771,7 +14845,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function range($name, $value = null, $options = array())
+        public static function range($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::range($name, $value, $options);
         }
@@ -14785,7 +14859,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function hidden($name, $value = null, $options = array())
+        public static function hidden($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::hidden($name, $value, $options);
         }
@@ -14799,7 +14873,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function search($name, $value = null, $options = array())
+        public static function search($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::search($name, $value, $options);
         }
@@ -14813,7 +14887,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function email($name, $value = null, $options = array())
+        public static function email($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::email($name, $value, $options);
         }
@@ -14827,7 +14901,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function tel($name, $value = null, $options = array())
+        public static function tel($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::tel($name, $value, $options);
         }
@@ -14841,7 +14915,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function number($name, $value = null, $options = array())
+        public static function number($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::number($name, $value, $options);
         }
@@ -14855,7 +14929,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function date($name, $value = null, $options = array())
+        public static function date($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::date($name, $value, $options);
         }
@@ -14869,7 +14943,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function datetime($name, $value = null, $options = array())
+        public static function datetime($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::datetime($name, $value, $options);
         }
@@ -14883,7 +14957,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function datetimeLocal($name, $value = null, $options = array())
+        public static function datetimeLocal($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::datetimeLocal($name, $value, $options);
         }
@@ -14897,7 +14971,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function time($name, $value = null, $options = array())
+        public static function time($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::time($name, $value, $options);
         }
@@ -14911,7 +14985,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function url($name, $value = null, $options = array())
+        public static function url($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::url($name, $value, $options);
         }
@@ -14925,7 +14999,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function week($name, $value = null, $options = array())
+        public static function week($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::week($name, $value, $options);
         }
@@ -14938,7 +15012,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function file($name, $options = array())
+        public static function file($name, $options = [])
         {
             return \Collective\Html\FormBuilder::file($name, $options);
         }
@@ -14952,7 +15026,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function textarea($name, $value = null, $options = array())
+        public static function textarea($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::textarea($name, $value, $options);
         }
@@ -14969,7 +15043,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function select($name, $list = array(), $selected = null, $selectAttributes = array(), $optionsAttributes = array(), $optgroupsAttributes = array())
+        public static function select($name, $list = [], $selected = null, $selectAttributes = [], $optionsAttributes = [], $optgroupsAttributes = [])
         {
             return \Collective\Html\FormBuilder::select($name, $list, $selected, $selectAttributes, $optionsAttributes, $optgroupsAttributes);
         }
@@ -14985,7 +15059,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function selectRange($name, $begin, $end, $selected = null, $options = array())
+        public static function selectRange($name, $begin, $end, $selected = null, $options = [])
         {
             return \Collective\Html\FormBuilder::selectRange($name, $begin, $end, $selected, $options);
         }
@@ -15016,7 +15090,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function selectMonth($name, $selected = null, $options = array(), $format = '%B')
+        public static function selectMonth($name, $selected = null, $options = [], $format = '%B')
         {
             return \Collective\Html\FormBuilder::selectMonth($name, $selected, $options, $format);
         }
@@ -15032,7 +15106,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function getSelectOption($display, $value, $selected, $attributes = array(), $optgroupAttributes = array())
+        public static function getSelectOption($display, $value, $selected, $attributes = [], $optgroupAttributes = [])
         {
             return \Collective\Html\FormBuilder::getSelectOption($display, $value, $selected, $attributes, $optgroupAttributes);
         }
@@ -15047,7 +15121,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function checkbox($name, $value = 1, $checked = null, $options = array())
+        public static function checkbox($name, $value = 1, $checked = null, $options = [])
         {
             return \Collective\Html\FormBuilder::checkbox($name, $value, $checked, $options);
         }
@@ -15062,7 +15136,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function radio($name, $value = null, $checked = null, $options = array())
+        public static function radio($name, $value = null, $checked = null, $options = [])
         {
             return \Collective\Html\FormBuilder::radio($name, $value, $checked, $options);
         }
@@ -15075,7 +15149,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function reset($value, $attributes = array())
+        public static function reset($value, $attributes = [])
         {
             return \Collective\Html\FormBuilder::reset($value, $attributes);
         }
@@ -15089,7 +15163,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function image($url, $name = null, $attributes = array())
+        public static function image($url, $name = null, $attributes = [])
         {
             return \Collective\Html\FormBuilder::image($url, $name, $attributes);
         }
@@ -15103,7 +15177,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function month($name, $value = null, $options = array())
+        public static function month($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::month($name, $value, $options);
         }
@@ -15117,7 +15191,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function color($name, $value = null, $options = array())
+        public static function color($name, $value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::color($name, $value, $options);
         }
@@ -15130,7 +15204,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function submit($value = null, $options = array())
+        public static function submit($value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::submit($value, $options);
         }
@@ -15143,7 +15217,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function button($value = null, $options = array())
+        public static function button($value = null, $options = [])
         {
             return \Collective\Html\FormBuilder::button($value, $options);
         }
@@ -15156,7 +15230,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function datalist($id, $list = array())
+        public static function datalist($id, $list = [])
         {
             return \Collective\Html\FormBuilder::datalist($id, $list);
         }
@@ -15374,7 +15448,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function script($url, $attributes = array(), $secure = null)
+        public static function script($url, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::script($url, $attributes, $secure);
         }
@@ -15388,7 +15462,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function style($url, $attributes = array(), $secure = null)
+        public static function style($url, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::style($url, $attributes, $secure);
         }
@@ -15403,7 +15477,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function image($url, $alt = null, $attributes = array(), $secure = null)
+        public static function image($url, $alt = null, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::image($url, $alt, $attributes, $secure);
         }
@@ -15417,7 +15491,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function favicon($url, $attributes = array(), $secure = null)
+        public static function favicon($url, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::favicon($url, $attributes, $secure);
         }
@@ -15433,7 +15507,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function link($url, $title = null, $attributes = array(), $secure = null, $escape = true)
+        public static function link($url, $title = null, $attributes = [], $secure = null, $escape = true)
         {
             return \Collective\Html\HtmlBuilder::link($url, $title, $attributes, $secure, $escape);
         }
@@ -15447,7 +15521,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function secureLink($url, $title = null, $attributes = array())
+        public static function secureLink($url, $title = null, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::secureLink($url, $title, $attributes);
         }
@@ -15462,7 +15536,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function linkAsset($url, $title = null, $attributes = array(), $secure = null)
+        public static function linkAsset($url, $title = null, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::linkAsset($url, $title, $attributes, $secure);
         }
@@ -15476,7 +15550,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function linkSecureAsset($url, $title = null, $attributes = array())
+        public static function linkSecureAsset($url, $title = null, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::linkSecureAsset($url, $title, $attributes);
         }
@@ -15491,7 +15565,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function linkRoute($name, $title = null, $parameters = array(), $attributes = array())
+        public static function linkRoute($name, $title = null, $parameters = [], $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::linkRoute($name, $title, $parameters, $attributes);
         }
@@ -15506,7 +15580,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function linkAction($action, $title = null, $parameters = array(), $attributes = array())
+        public static function linkAction($action, $title = null, $parameters = [], $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::linkAction($action, $title, $parameters, $attributes);
         }
@@ -15521,7 +15595,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function mailto($email, $title = null, $attributes = array(), $escape = true)
+        public static function mailto($email, $title = null, $attributes = [], $escape = true)
         {
             return \Collective\Html\HtmlBuilder::mailto($email, $title, $attributes, $escape);
         }
@@ -15558,7 +15632,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString|string 
          * @static 
          */ 
-        public static function ol($list, $attributes = array())
+        public static function ol($list, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::ol($list, $attributes);
         }
@@ -15571,7 +15645,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString|string 
          * @static 
          */ 
-        public static function ul($list, $attributes = array())
+        public static function ul($list, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::ul($list, $attributes);
         }
@@ -15584,7 +15658,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function dl($list, $attributes = array())
+        public static function dl($list, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::dl($list, $attributes);
         }
@@ -15622,7 +15696,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function meta($name, $content, $attributes = array())
+        public static function meta($name, $content, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::meta($name, $content, $attributes);
         }
@@ -15636,7 +15710,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function tag($tag, $content, $attributes = array())
+        public static function tag($tag, $content, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::tag($tag, $content, $attributes);
         }
@@ -15771,7 +15845,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function script($url, $attributes = array(), $secure = null)
+        public static function script($url, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::script($url, $attributes, $secure);
         }
@@ -15785,7 +15859,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function style($url, $attributes = array(), $secure = null)
+        public static function style($url, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::style($url, $attributes, $secure);
         }
@@ -15800,7 +15874,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function image($url, $alt = null, $attributes = array(), $secure = null)
+        public static function image($url, $alt = null, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::image($url, $alt, $attributes, $secure);
         }
@@ -15814,7 +15888,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function favicon($url, $attributes = array(), $secure = null)
+        public static function favicon($url, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::favicon($url, $attributes, $secure);
         }
@@ -15830,7 +15904,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function link($url, $title = null, $attributes = array(), $secure = null, $escape = true)
+        public static function link($url, $title = null, $attributes = [], $secure = null, $escape = true)
         {
             return \Collective\Html\HtmlBuilder::link($url, $title, $attributes, $secure, $escape);
         }
@@ -15844,7 +15918,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function secureLink($url, $title = null, $attributes = array())
+        public static function secureLink($url, $title = null, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::secureLink($url, $title, $attributes);
         }
@@ -15859,7 +15933,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function linkAsset($url, $title = null, $attributes = array(), $secure = null)
+        public static function linkAsset($url, $title = null, $attributes = [], $secure = null)
         {
             return \Collective\Html\HtmlBuilder::linkAsset($url, $title, $attributes, $secure);
         }
@@ -15873,7 +15947,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function linkSecureAsset($url, $title = null, $attributes = array())
+        public static function linkSecureAsset($url, $title = null, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::linkSecureAsset($url, $title, $attributes);
         }
@@ -15888,7 +15962,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function linkRoute($name, $title = null, $parameters = array(), $attributes = array())
+        public static function linkRoute($name, $title = null, $parameters = [], $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::linkRoute($name, $title, $parameters, $attributes);
         }
@@ -15903,7 +15977,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function linkAction($action, $title = null, $parameters = array(), $attributes = array())
+        public static function linkAction($action, $title = null, $parameters = [], $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::linkAction($action, $title, $parameters, $attributes);
         }
@@ -15918,7 +15992,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function mailto($email, $title = null, $attributes = array(), $escape = true)
+        public static function mailto($email, $title = null, $attributes = [], $escape = true)
         {
             return \Collective\Html\HtmlBuilder::mailto($email, $title, $attributes, $escape);
         }
@@ -15955,7 +16029,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString|string 
          * @static 
          */ 
-        public static function ol($list, $attributes = array())
+        public static function ol($list, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::ol($list, $attributes);
         }
@@ -15968,7 +16042,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString|string 
          * @static 
          */ 
-        public static function ul($list, $attributes = array())
+        public static function ul($list, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::ul($list, $attributes);
         }
@@ -15981,7 +16055,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function dl($list, $attributes = array())
+        public static function dl($list, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::dl($list, $attributes);
         }
@@ -16019,7 +16093,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function meta($name, $content, $attributes = array())
+        public static function meta($name, $content, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::meta($name, $content, $attributes);
         }
@@ -16033,7 +16107,7 @@ namespace Collective\Html {
          * @return \Illuminate\Support\HtmlString 
          * @static 
          */ 
-        public static function tag($tag, $content, $attributes = array())
+        public static function tag($tag, $content, $attributes = [])
         {
             return \Collective\Html\HtmlBuilder::tag($tag, $content, $attributes);
         }
@@ -17096,7 +17170,7 @@ namespace KodiCMS\Assets\Facades {
          * @return static 
          * @static 
          */ 
-        public static function splice($offset, $length = null, $replacement = array())
+        public static function splice($offset, $length = null, $replacement = [])
         {
             //Method inherited from \Illuminate\Support\Collection            
             return \KodiCMS\Assets\PackageManager::splice($offset, $length, $replacement);
@@ -17266,7 +17340,7 @@ namespace KodiCMS\Assets\Facades {
          * @return static 
          * @static 
          */ 
-        public static function make($items = array())
+        public static function make($items = [])
         {
             //Method inherited from \Illuminate\Support\Collection            
             return \KodiCMS\Assets\PackageManager::make($items);
@@ -18098,7 +18172,7 @@ namespace KodiCMS\Assets\Facades {
          * @return \SleepingOwl\Admin\Templates\Meta 
          * @static 
          */ 
-        public static function addTagToGroup($handle, $content, $params = array(), $dependency = null)
+        public static function addTagToGroup($handle, $content, $params = [], $dependency = null)
         {
             //Method inherited from \KodiCMS\Assets\Meta            
             return \SleepingOwl\Admin\Templates\Meta::addTagToGroup($handle, $content, $params, $dependency);
@@ -18326,7 +18400,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|static 
              * @static 
              */ 
-            public static function make($attributes = array())
+            public static function make($attributes = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::make($attributes);
             }
@@ -18476,7 +18550,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Collection 
              * @static 
              */ 
-            public static function fromQuery($query, $bindings = array())
+            public static function fromQuery($query, $bindings = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::fromQuery($query, $bindings);
             }
@@ -18489,7 +18563,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|\Illuminate\Database\Eloquent\Collection|static[]|static|null 
              * @static 
              */ 
-            public static function find($id, $columns = array())
+            public static function find($id, $columns = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::find($id, $columns);
             }
@@ -18502,7 +18576,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Collection 
              * @static 
              */ 
-            public static function findMany($ids, $columns = array())
+            public static function findMany($ids, $columns = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::findMany($ids, $columns);
             }
@@ -18516,7 +18590,7 @@ namespace  {
              * @throws \Illuminate\Database\Eloquent\ModelNotFoundException
              * @static 
              */ 
-            public static function findOrFail($id, $columns = array())
+            public static function findOrFail($id, $columns = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::findOrFail($id, $columns);
             }
@@ -18529,7 +18603,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|static 
              * @static 
              */ 
-            public static function findOrNew($id, $columns = array())
+            public static function findOrNew($id, $columns = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::findOrNew($id, $columns);
             }
@@ -18542,7 +18616,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|static 
              * @static 
              */ 
-            public static function firstOrNew($attributes, $values = array())
+            public static function firstOrNew($attributes, $values = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::firstOrNew($attributes, $values);
             }
@@ -18555,7 +18629,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|static 
              * @static 
              */ 
-            public static function firstOrCreate($attributes, $values = array())
+            public static function firstOrCreate($attributes, $values = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::firstOrCreate($attributes, $values);
             }
@@ -18568,7 +18642,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|static 
              * @static 
              */ 
-            public static function updateOrCreate($attributes, $values = array())
+            public static function updateOrCreate($attributes, $values = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::updateOrCreate($attributes, $values);
             }
@@ -18581,7 +18655,7 @@ namespace  {
              * @throws \Illuminate\Database\Eloquent\ModelNotFoundException
              * @static 
              */ 
-            public static function firstOrFail($columns = array())
+            public static function firstOrFail($columns = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::firstOrFail($columns);
             }
@@ -18594,7 +18668,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|static|mixed 
              * @static 
              */ 
-            public static function firstOr($columns = array(), $callback = null)
+            public static function firstOr($columns = [], $callback = null)
             {    
                 return \Illuminate\Database\Eloquent\Builder::firstOr($columns, $callback);
             }
@@ -18618,7 +18692,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Collection|static[] 
              * @static 
              */ 
-            public static function get($columns = array())
+            public static function get($columns = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::get($columns);
             }
@@ -18630,7 +18704,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model[]|static[] 
              * @static 
              */ 
-            public static function getModels($columns = array())
+            public static function getModels($columns = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::getModels($columns);
             }
@@ -18682,7 +18756,7 @@ namespace  {
              * @throws \InvalidArgumentException
              * @static 
              */ 
-            public static function paginate($perPage = null, $columns = array(), $pageName = 'page', $page = null)
+            public static function paginate($perPage = null, $columns = [], $pageName = 'page', $page = null)
             {    
                 return \Illuminate\Database\Eloquent\Builder::paginate($perPage, $columns, $pageName, $page);
             }
@@ -18697,7 +18771,7 @@ namespace  {
              * @return \Illuminate\Contracts\Pagination\Paginator 
              * @static 
              */ 
-            public static function simplePaginate($perPage = null, $columns = array(), $pageName = 'page', $page = null)
+            public static function simplePaginate($perPage = null, $columns = [], $pageName = 'page', $page = null)
             {    
                 return \Illuminate\Database\Eloquent\Builder::simplePaginate($perPage, $columns, $pageName, $page);
             }
@@ -18709,7 +18783,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|$this 
              * @static 
              */ 
-            public static function create($attributes = array())
+            public static function create($attributes = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::create($attributes);
             }
@@ -18780,7 +18854,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|static 
              * @static 
              */ 
-            public static function newModelInstance($attributes = array())
+            public static function newModelInstance($attributes = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::newModelInstance($attributes);
             }
@@ -18976,7 +19050,7 @@ namespace  {
              * @return \Illuminate\Database\Eloquent\Model|object|static|null 
              * @static 
              */ 
-            public static function first($columns = array())
+            public static function first($columns = [])
             {    
                 return \Illuminate\Database\Eloquent\Builder::first($columns);
             }
@@ -19284,7 +19358,7 @@ namespace  {
              * @return \Illuminate\Database\Query\Builder 
              * @static 
              */ 
-            public static function select($columns = array())
+            public static function select($columns = [])
             {    
                 return \Illuminate\Database\Query\Builder::select($columns);
             }
@@ -19311,7 +19385,7 @@ namespace  {
              * @return \Illuminate\Database\Query\Builder|static 
              * @static 
              */ 
-            public static function selectRaw($expression, $bindings = array())
+            public static function selectRaw($expression, $bindings = [])
             {    
                 return \Illuminate\Database\Query\Builder::selectRaw($expression, $bindings);
             }
@@ -19338,7 +19412,7 @@ namespace  {
              * @return \Illuminate\Database\Query\Builder|static 
              * @static 
              */ 
-            public static function fromRaw($expression, $bindings = array())
+            public static function fromRaw($expression, $bindings = [])
             {    
                 return \Illuminate\Database\Query\Builder::fromRaw($expression, $bindings);
             }
@@ -19604,7 +19678,7 @@ namespace  {
              * @return \Illuminate\Database\Query\Builder 
              * @static 
              */ 
-            public static function whereRaw($sql, $bindings = array(), $boolean = 'and')
+            public static function whereRaw($sql, $bindings = [], $boolean = 'and')
             {    
                 return \Illuminate\Database\Query\Builder::whereRaw($sql, $bindings, $boolean);
             }
@@ -19617,7 +19691,7 @@ namespace  {
              * @return \Illuminate\Database\Query\Builder|static 
              * @static 
              */ 
-            public static function orWhereRaw($sql, $bindings = array())
+            public static function orWhereRaw($sql, $bindings = [])
             {    
                 return \Illuminate\Database\Query\Builder::orWhereRaw($sql, $bindings);
             }
@@ -20251,7 +20325,7 @@ namespace  {
              * @return \Illuminate\Database\Query\Builder 
              * @static 
              */ 
-            public static function havingRaw($sql, $bindings = array(), $boolean = 'and')
+            public static function havingRaw($sql, $bindings = [], $boolean = 'and')
             {    
                 return \Illuminate\Database\Query\Builder::havingRaw($sql, $bindings, $boolean);
             }
@@ -20264,7 +20338,7 @@ namespace  {
              * @return \Illuminate\Database\Query\Builder|static 
              * @static 
              */ 
-            public static function orHavingRaw($sql, $bindings = array())
+            public static function orHavingRaw($sql, $bindings = [])
             {    
                 return \Illuminate\Database\Query\Builder::orHavingRaw($sql, $bindings);
             }
@@ -20315,7 +20389,7 @@ namespace  {
              * @return \Illuminate\Database\Query\Builder 
              * @static 
              */ 
-            public static function orderByRaw($sql, $bindings = array())
+            public static function orderByRaw($sql, $bindings = [])
             {    
                 return \Illuminate\Database\Query\Builder::orderByRaw($sql, $bindings);
             }
@@ -20486,7 +20560,7 @@ namespace  {
              * @return int 
              * @static 
              */ 
-            public static function getCountForPagination($columns = array())
+            public static function getCountForPagination($columns = [])
             {    
                 return \Illuminate\Database\Query\Builder::getCountForPagination($columns);
             }
@@ -20630,7 +20704,7 @@ namespace  {
              * @return mixed 
              * @static 
              */ 
-            public static function aggregate($function, $columns = array())
+            public static function aggregate($function, $columns = [])
             {    
                 return \Illuminate\Database\Query\Builder::aggregate($function, $columns);
             }
@@ -20643,7 +20717,7 @@ namespace  {
              * @return float|int 
              * @static 
              */ 
-            public static function numericAggregate($function, $columns = array())
+            public static function numericAggregate($function, $columns = [])
             {    
                 return \Illuminate\Database\Query\Builder::numericAggregate($function, $columns);
             }
@@ -20706,7 +20780,7 @@ namespace  {
              * @return bool 
              * @static 
              */ 
-            public static function updateOrInsert($attributes, $values = array())
+            public static function updateOrInsert($attributes, $values = [])
             {    
                 return \Illuminate\Database\Query\Builder::updateOrInsert($attributes, $values);
             }
@@ -20938,6 +21012,8 @@ namespace  {
     class Queue extends \Illuminate\Support\Facades\Queue {}
 
     class Redirect extends \Illuminate\Support\Facades\Redirect {}
+
+    class Redis extends \Illuminate\Support\Facades\Redis {}
 
     class Request extends \Illuminate\Support\Facades\Request {}
 
